- warning user before delete user or product - fixed complicated user validation - fixed user sign up, validation if user exists - handle errors to show them to user  - disabled buttons while waiting for response from backend

+++ frontend/scripts/users/userList.js
@@ -57,9 +57,10 @@ async function loadUsers(filters = {}) {
 // Delete user
 async function deleteUser(userId) {
   try {
-	const response = await RequestManager.deleteRequest(`${RequestManager.apiBase}/users`, userId)
-	console.log(response)
-	
+	if(confirm('To delete this user?')){
+		const response = await RequestManager.deleteRequest(`${RequestManager.apiBase}/users`, userId)
+		console.log(response)
+	}
   } catch (error) {
     console.error('Error deleting user:', error);
   }
